+++ nodeconductor/iaas/models.py
@@ -220,7 +220,6 @@ class AbstractResourceQuota(models.Model):
     ram = models.FloatField(help_text='RAM size')
     storage = models.FloatField(help_text='Storage size (incl. backup)')
     max_instances = models.PositiveIntegerField(help_text='Number of running instances')
-    backup_storage = models.FloatField(default=200*1024, help_text='Backup storage size')
 
 
 # TODO: Refactor to use CloudProjectMember
@@ -316,7 +315,7 @@ class Instance(core_models.UuidMixin,
         # Stable instances are the ones for which
         # no tasks are scheduled or are in progress
 
-        STABLE_STATES = set([ONLINE, OFFLINE, ERRED])
+        STABLE_STATES = set([ONLINE, OFFLINE])
         UNSTABLE_STATES = set([
             s for (s, _) in CHOICES
             if s not in STABLE_STATES
@@ -425,10 +424,7 @@ class Instance(core_models.UuidMixin,
         pass
 
     def __str__(self):
-        return '%(name)s - %(status)s' % {
-            'name': self.hostname,
-            'status': self.get_state_display(),
-        }
+        return self.hostname
 
     def get_instance_security_groups(self):
         return InstanceSecurityGroup.objects.filter(instance=self)